✨ 홈탭 프로필 카드 전면/후면 구현
홈탭의 펫 정보 전면/후면 카드 구현
