+++ ttakkeun/ttakkeun/Components/Home/HomeFrontCard.swift
@@ -35,9 +35,9 @@ struct HomeFrontCard: View {
             })
             .padding(.horizontal, 24)
             .padding(.top, 13)
-           
             
-            profileImageGroup
+            
+            ProfileImageCard(imageUrl: viewModel.profileData?.imageUrl ?? "")
             
             profileInfor
             
@@ -59,51 +59,7 @@ struct HomeFrontCard: View {
         .frame(width: 354, height: 240)
         .background(Color.white.opacity(0.5))
         .clipShape(.rect(cornerRadius: 20))
-    }
-    
-    // MARK: - ProfileImage
-    
-    /// 프로필 이미지
-    private var profileImageGroup: some View {
-        ZStack(alignment: .bottomTrailing, content: {
-            profileImage
-            editProfileImage
-        })
-    }
-    
-    /// 프로필 이미지 뷰 빌더
-    @ViewBuilder
-    private var profileImage: some View {
-        if let url = URL(string: viewModel.profileData?.imageUrl ?? "") {
-            KFImage(url)
-                .placeholder {
-                    ProgressView()
-                        .frame(width: 100, height: 100)
-                }.retry(maxCount: 2, interval: .seconds(2))
-                .onFailure { _ in
-                    print("프로필 이미지 캐시 오류")
-                }
-                .resizable()
-                .frame(width: 140, height: 140)
-                .clipShape(Circle())
-        }
-    }
-    
-    /// 프로필 사진 편집 버튼
-    private var editProfileImage: some View {
-        Button(action: {
-            print("hello")
-        }, label: {
-            ZStack(alignment: .center, content: {
-                Circle()
-                    .fill(Color.editColor_Color)
-                    .frame(width: 30, height: 30)
-                Icon.editProfile.image
-                    .resizable()
-                    .frame(width: 14, height: 14)
-            })
-            .padding([.vertical, .horizontal], 5)
-        })
+        .shadow02()
     }
